requirements.txt
python-dotenv
gspread
fastapi
quickemailverification
paramiko
pytz